fix alignment in readSourceFileToEndAlloc
Previously, it would return an align(2) slice as provided by
toOwnedSliceSentinel, which was problematic as the callee owns the
memory and would mistakingly free it as an align(1) slice.

The array list was originally changed to be align(2) to avoid
reallocation when converting UTF16LE to UTF8, however this changes it
back to align(1) and instead adds a function to std.unicode to convert
unaligned codepoints.